+++ src/nocrypto.mli
@@ -25,7 +25,7 @@
     It is here only temporary, until we find it a proper home. *)
 module Base64 : sig
   val encode : Cstruct.t -> Cstruct.t
-  val decode : Cstruct.t -> Cstruct.t
+  val decode : Cstruct.t -> Cstruct.t option
   val is_base64_char : char -> bool
 end